Rising-Phoenix/SoverignAI/translator/requirements.txt
google-cloud-translate==3.15.5
python-dotenv==1.0.0 